Update tests and snapshots, including unrolled materials
