+++ app/assets/javascripts/reader_profile_january/FAndPMaterials.story.js
@@ -1,15 +1,14 @@
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import FAndPMaterials, {UnrolledForTest} from './FAndPMaterials';
+import FAndPMaterials from './FAndPMaterials';
+import {UnrolledForTest} from './FAndPMaterials.test';
 
 
 function renderLevel(rawLevelText) {
   return <div style={styles.container}><FAndPMaterials rawLevelText={rawLevelText} /></div>;
 }
 
-console.log('remove unrolled');
 storiesOf('reader_profile_january/FAndPMaterials', module) // eslint-disable-line no-undef
-  .add('unrolled', () => <UnrolledForTest levelStyle={styles.container} />)
   .add('default', () => (
     <div>
       {renderLevel("A")}
@@ -39,7 +38,8 @@ storiesOf('reader_profile_january/FAndPMaterials', module) // eslint-disable-lin
       {renderLevel("Y")}
       {renderLevel("Z")}
     </div>
-  ));
+  ))
+  .add('unrolled', () => <UnrolledForTest levelStyle={styles.container} />);
 
 const styles = {
   container: {
